+++ sdk/data/aztables/transaction_batch_test.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
-	"github.com/Azure/azure-sdk-for-go/sdk/internal/uuid"
 	"github.com/stretchr/testify/require"
 )
 
@@ -28,7 +27,7 @@ func TestBatchAdd(t *testing.T) {
 			entitiesToCreate := createComplexEntities(10, "partition")
 			var batch []TransactionAction
 
-			for _, e := range *entitiesToCreate {
+			for _, e := range entitiesToCreate {
 				marshalled, err := json.Marshal(e)
 				require.NoError(t, err)
 				batch = append(batch, TransactionAction{ActionType: TransactionTypeAdd, Entity: marshalled})
@@ -37,8 +36,8 @@ func TestBatchAdd(t *testing.T) {
 			resp, err := client.SubmitTransaction(ctx, batch, nil)
 
 			require.NoError(t, err)
-			for i := 0; i < len(*resp.TransactionResponses); i++ {
-				r := (*resp.TransactionResponses)[i]
+			for i := 0; i < len(resp.TransactionResponses); i++ {
+				r := (resp.TransactionResponses)[i]
 				require.Equal(t, r.StatusCode, http.StatusNoContent)
 			}
 
@@ -68,7 +67,7 @@ func TestBatchInsert(t *testing.T) {
 			entitiesToCreate := createComplexEntities(1, "partition")
 			var batch []TransactionAction
 
-			for _, e := range *entitiesToCreate {
+			for _, e := range entitiesToCreate {
 				marshalled, err := json.Marshal(e)
 				require.NoError(t, err)
 				batch = append(
@@ -82,8 +81,8 @@ func TestBatchInsert(t *testing.T) {
 
 			resp, err := client.SubmitTransaction(ctx, batch, nil)
 			require.NoError(t, err)
-			for i := 1; i < len(*resp.TransactionResponses); i++ {
-				r := (*resp.TransactionResponses)[i]
+			for i := 1; i < len(resp.TransactionResponses); i++ {
+				r := (resp.TransactionResponses)[i]
 				require.Equal(t, r.StatusCode, http.StatusNoContent)
 			}
 
@@ -112,7 +111,7 @@ func TestBatchMixed(t *testing.T) {
 			entitiesToCreate := createComplexEntities(5, "partition")
 			var batch []TransactionAction
 
-			for _, e := range *entitiesToCreate {
+			for _, e := range entitiesToCreate {
 				marshalled, err := json.Marshal(e)
 				require.NoError(t, err)
 				batch = append(batch, TransactionAction{
@@ -123,8 +122,8 @@ func TestBatchMixed(t *testing.T) {
 
 			resp, err := client.SubmitTransaction(ctx, batch, nil)
 			require.NoError(t, err)
-			for i := 0; i < len(*resp.TransactionResponses); i++ {
-				r := (*resp.TransactionResponses)[i]
+			for i := 0; i < len(resp.TransactionResponses); i++ {
+				r := (resp.TransactionResponses)[i]
 				require.Equal(t, http.StatusNoContent, r.StatusCode)
 			}
 
@@ -147,13 +146,13 @@ func TestBatchMixed(t *testing.T) {
 			mergeProp := "MergeProperty"
 			val := "foo"
 			var mergeEntity = map[string]interface{}{
-				partitionKey: (*entitiesToCreate)[0].PartitionKey,
-				rowKey:       (*entitiesToCreate)[0].RowKey,
+				partitionKey: (entitiesToCreate)[0].PartitionKey,
+				rowKey:       (entitiesToCreate)[0].RowKey,
 				mergeProp:    val,
 			}
 			marshalledMergeEntity, err := json.Marshal(mergeEntity)
 			require.NoError(t, err)
-			etag := azcore.ETag((*resp.TransactionResponses)[0].Header.Get(etag))
+			etag := azcore.ETag((resp.TransactionResponses)[0].Header.Get(etag))
 			batch2 = append(batch2, TransactionAction{
 				ActionType: TransactionTypeUpdateMerge,
 				Entity:     marshalledMergeEntity,
@@ -161,15 +160,15 @@ func TestBatchMixed(t *testing.T) {
 			})
 
 			// create a delete action for the second added entity
-			marshalledSecondEntity, err := json.Marshal((*entitiesToCreate)[1])
+			marshalledSecondEntity, err := json.Marshal((entitiesToCreate)[1])
 			require.NoError(t, err)
 			batch2 = append(batch2, TransactionAction{ActionType: TransactionTypeDelete, Entity: marshalledSecondEntity})
 
 			// create an insert action to replace the third added entity with a new value
 			replaceProp := "ReplaceProperty"
 			var replaceProperties = map[string]interface{}{
-				partitionKey: (*entitiesToCreate)[2].PartitionKey,
-				rowKey:       (*entitiesToCreate)[2].RowKey,
+				partitionKey: (entitiesToCreate)[2].PartitionKey,
+				rowKey:       (entitiesToCreate)[2].RowKey,
 				replaceProp:  val,
 			}
 			marshalledThirdEntity, err := json.Marshal(replaceProperties)
@@ -177,9 +176,9 @@ func TestBatchMixed(t *testing.T) {
 			batch2 = append(batch2, TransactionAction{ActionType: TransactionTypeInsertReplace, Entity: marshalledThirdEntity})
 
 			// Add the remaining 2 entities.
-			marshalled4thEntity, err := json.Marshal((*entitiesToCreate)[3])
+			marshalled4thEntity, err := json.Marshal((entitiesToCreate)[3])
 			require.NoError(t, err)
-			marshalled5thEntity, err := json.Marshal((*entitiesToCreate)[4])
+			marshalled5thEntity, err := json.Marshal((entitiesToCreate)[4])
 			require.NoError(t, err)
 			batch2 = append(batch2, TransactionAction{ActionType: TransactionTypeUpdateMerge, Entity: marshalled4thEntity})
 			batch2 = append(batch2, TransactionAction{ActionType: TransactionTypeInsertMerge, Entity: marshalled5thEntity})
@@ -187,8 +186,8 @@ func TestBatchMixed(t *testing.T) {
 			resp, err = client.SubmitTransaction(ctx, batch2, nil)
 			require.NoError(t, err)
 
-			for i := 0; i < len(*resp.TransactionResponses); i++ {
-				r := (*resp.TransactionResponses)[i]
+			for i := 0; i < len(resp.TransactionResponses); i++ {
+				r := (resp.TransactionResponses)[i]
 				require.Equal(t, http.StatusNoContent, r.StatusCode)
 
 			}
@@ -224,32 +223,70 @@ func TestBatchError(t *testing.T) {
 			// Create the batch.
 			var batch []TransactionAction
 
-			u1, err := uuid.New()
-			require.NoError(t, err)
-			u2, err := uuid.New()
-			require.NoError(t, err)
-
 			// Sending an empty batch throws.
-			_, err = client.submitTransactionInternal(ctx, &batch, u1, u2, nil)
-			require.NotNil(t, err)
+			_, err = client.SubmitTransaction(ctx, batch, nil)
+			require.Error(t, err)
 			require.Equal(t, errEmptyTransaction.Error(), err.Error())
 
 			// Add the last entity to the table prior to adding it as part of the batch to cause a batch failure.
-			marshalledFinalEntity, err := json.Marshal((*entitiesToCreate)[2])
+			marshalledFinalEntity, err := json.Marshal((entitiesToCreate)[2])
 			require.NoError(t, err)
 			_, err = client.AddEntity(ctx, marshalledFinalEntity, nil)
 			require.NoError(t, err)
 
 			// Add the entities to the batch
-			for _, e := range *entitiesToCreate {
+			for _, e := range entitiesToCreate {
 				marshalledEntity, err := json.Marshal(e)
 				require.NoError(t, err)
 				batch = append(batch, TransactionAction{ActionType: TransactionTypeAdd, Entity: marshalledEntity})
 			}
 
 			_, err = client.SubmitTransaction(ctx, batch, nil)
-			require.NotNil(t, err)
-			require.Contains(t, err.Error(), "EntityAlreadyExists")
+			require.Error(t, err)
+		})
+	}
+}
+
+func TestBatchErrorHandleResponse(t *testing.T) {
+	for _, service := range services {
+		t.Run(fmt.Sprintf("%v_%v", t.Name(), service), func(t *testing.T) {
+			client, deleteAndStop := initClientTest(t, service, true)
+			defer deleteAndStop()
+			err := recording.SetBodilessMatcher(t, nil)
+			require.NoError(t, err)
+			err = recording.AddGeneralRegexSanitizer("batch_00000000-0000-0000-0000-000000000000", "batch_[0-9A-Fa-f]{8}[-]([0-9A-Fa-f]{4}[-]?){3}[0-9a-fA-F]{12}", nil)
+			require.NoError(t, err)
+
+			entitiesToCreate := createComplexEntities(3, "partition")
+
+			// Create the batch.
+			var batch []TransactionAction
+
+			for _, e := range entitiesToCreate {
+				marshalled, err := json.Marshal(e)
+				require.NoError(t, err)
+				batch = append(batch, TransactionAction{
+					ActionType: TransactionTypeAdd,
+					Entity:     marshalled,
+				})
+			}
+
+			// Add the first entity a second type
+			marshalled, err := json.Marshal(entitiesToCreate[0])
+			require.NoError(t, err)
+			batch = append(batch, TransactionAction{
+				ActionType: TransactionTypeAdd,
+				Entity:     marshalled,
+			})
+
+			// Sending a batch with two entities on the same row returns an error
+			resp, err := client.SubmitTransaction(ctx, batch, nil)
+			require.Error(t, err)
+			if service == "storage" {
+				require.NotNil(t, resp.RawResponse)
+			} else {
+				require.Nil(t, resp.RawResponse)
+			}
 		})
 	}
 }
@@ -308,8 +345,8 @@ func TestBatchComplex(t *testing.T) {
 
 			resp, err := client.SubmitTransaction(ctx, batch, nil)
 			require.NoError(t, err)
-			for i := 0; i < len(*resp.TransactionResponses); i++ {
-				r := (*resp.TransactionResponses)[i]
+			for i := 0; i < len(resp.TransactionResponses); i++ {
+				r := (resp.TransactionResponses)[i]
 				require.Equal(t, http.StatusNoContent, r.StatusCode)
 			}
 
@@ -340,8 +377,8 @@ func TestBatchComplex(t *testing.T) {
 
 			resp, err = client.SubmitTransaction(ctx, batch2, nil)
 			require.NoError(t, err)
-			for i := 0; i < len(*resp.TransactionResponses); i++ {
-				r := (*resp.TransactionResponses)[i]
+			for i := 0; i < len(resp.TransactionResponses); i++ {
+				r := (resp.TransactionResponses)[i]
 				require.Equal(t, http.StatusNoContent, r.StatusCode)
 			}