+++ sdk/data/aztables/CHANGELOG.md
@@ -11,6 +11,7 @@
 * Changed the `ListTablesPager` and `ListEntitiesPager` to structs.
 * Renamed the `ResponseProperties` type to `TableProperties`.
 * Returns `Client` and `ServiceClient` by value instead of by reference.
+* Removing `ContentType` from the `TransactionResponse` struct.
 
 ### Bugs Fixed